Move performance tests out of package:devtools_testing
