+++ packages/devtools_app/test/performance_model_test.dart
@@ -8,9 +8,10 @@ import 'package:devtools_app/src/profiler/cpu_profile_model.dart';
 import 'package:devtools_app/src/trace_event.dart';
 import 'package:devtools_app/src/utils.dart';
 import 'package:devtools_testing/support/cpu_profile_test_data.dart';
-import 'package:devtools_testing/support/performance_test_data.dart';
 import 'package:devtools_testing/support/test_utils.dart';
 import 'package:flutter_test/flutter_test.dart';
+
+import 'support/performance_test_data.dart';
 
 void main() {
   group('PerformanceData', () {
