+++ packages/devtools_app/test/performance_screen_test.dart
@@ -12,13 +12,13 @@ import 'package:devtools_app/src/performance/performance_screen.dart';
 import 'package:devtools_app/src/performance/timeline_flame_chart.dart';
 import 'package:devtools_app/src/service_manager.dart';
 import 'package:devtools_app/src/split.dart';
-import 'package:devtools_testing/support/performance_test_data.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:vm_service/vm_service.dart' as vm_service;
 
 import 'support/mocks.dart';
+import 'support/performance_test_data.dart';
 import 'support/wrappers.dart';
 
 void main() {
