+++ packages/devtools_app/test/performance_utils_test.dart
@@ -3,8 +3,9 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app/src/performance/performance_utils.dart';
-import 'package:devtools_testing/support/performance_test_data.dart';
 import 'package:flutter_test/flutter_test.dart';
+
+import 'support/performance_test_data.dart';
 
 void main() {
   group('PerformanceUtils', () {
