+++ packages/devtools_app/test/timeline_flame_chart_test.dart
@@ -10,7 +10,6 @@ import 'package:devtools_app/src/performance/performance_controller.dart';
 import 'package:devtools_app/src/performance/performance_screen.dart';
 import 'package:devtools_app/src/performance/timeline_flame_chart.dart';
 import 'package:devtools_app/src/service_manager.dart';
-import 'package:devtools_testing/support/performance_test_data.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -18,6 +17,7 @@ import 'package:mockito/mockito.dart';
 import 'package:vm_service/vm_service.dart' as vm_service;
 
 import 'support/mocks.dart';
+import 'support/performance_test_data.dart';
 import 'support/wrappers.dart';
 
 void main() {
