+++ packages/devtools_app/test/timeline_processor_test.dart
@@ -7,10 +7,11 @@ import 'package:devtools_app/src/performance/performance_model.dart';
 import 'package:devtools_app/src/performance/timeline_event_processor.dart';
 import 'package:devtools_app/src/trace_event.dart';
 import 'package:devtools_app/src/utils.dart';
-import 'package:devtools_testing/support/performance_test_data.dart';
 import 'package:devtools_testing/support/test_utils.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
+
+import 'support/performance_test_data.dart';
 
 void main() {
   final originalGoldenUiEvent = goldenUiTimelineEvent.deepCopy();
